feat(metadata): Move all version fields to u64
The version field is represented as a JSON number, and is in other implementations (notable go-tuf) a 64 bits integer.

We've had incidents where we put the version of a target as the unix millis timestamp, which overflowed the 32 bits version of some implementations.

For the sake of compatibility, all versions are now 64 bits

+++ interop-tests/src/lib.rs
@@ -106,7 +106,7 @@ async fn update_root(
     repo: &mut FileSystemRepository<JsonPretty>,
     keys: &RoleKeys,
     root_signer: Option<&dyn PrivateKey>,
-    version: u32,
+    version: u64,
     consistent_snapshot: bool,
 ) {
     // Same expiration as go-tuf metadata generator.
@@ -148,7 +148,7 @@ async fn add_target(
 ) {
     // Same expiration as go-tuf metadata generator.
     let expiration = Utc.ymd(2100, 1, 1).and_hms(0, 0, 0);
-    let version: u32 = (step + 1).into();
+    let version: u64 = (step + 1).into();
 
     let mut targets_builder = TargetsMetadataBuilder::new()
         .expires(expiration)
